feat: markdown toolbar and preview extensions merged into phoenix markdown
