+++ src/extensions/default/Phoenix-Markdown/.gitignore
@@ -1,3 +1,4 @@
 samples/
 *.swp
 *.zip
+.DS_Store
