+++ src/extensions/default/Phoenix-Markdown/templates/panel.html
