--- src/extensions/default/Phoenix-markdown-toolbar/nls/root/strings.js
@@ -1,56 +1,0 @@
-/*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50 */
-/*global define */
-
-define({
-    "CANCEL" : "Cancel",
-    "OK"     : "OK",
-    
-    "DESCRIPTION_SHOW_ON_STARTUP" : "Show Markdown bar on startup",
-    "DESCRIPTION_MAX_LINE_LENGTH" : "Maximum line length for reflow",
-
-    "MENU_TOOLBAR"     : "Markdown Toolbar",
-
-    "BUTTON_H1"        : "H1",
-    "BUTTON_H2"        : "H2",
-    "BUTTON_H3"        : "H3",
-    "BUTTON_H4"        : "H4",
-    "BUTTON_H5"        : "H5",
-    "BUTTON_H6"        : "H6",
-    "BUTTON_BOLD"      : "B",
-    "BUTTON_ITALIC"    : "I",
-    "BUTTON_STRIKE"    : "S",
-    "BUTTON_CODE"      : "ab",
-
-    "HINT_H1"          : "Heading 1",
-    "HINT_H2"          : "Heading 2",
-    "HINT_H3"          : "Heading 3",
-    "HINT_H4"          : "Heading 4",
-    "HINT_H5"          : "Heading 5",
-    "HINT_H6"          : "Heading 6",
-    "HINT_BOLD"        : "Bold",
-    "HINT_ITALIC"      : "Italic",
-    "HINT_STRIKE"      : "Strikethrough",
-    "HINT_CODE"        : "Inline Code",
-    "HINT_BULLET"      : "Bulleted List",
-    "HINT_NUMBERED"    : "Numbered List",
-    "HINT_QUOTE"       : "Quoted Text",
-    "HINT_CODEBLOCK"   : "Code Block",
-    "HINT_IMAGE"       : "Image",
-    "HINT_LINK"        : "Link",
-    "HINT_PARAGRAPH"   : "Separate Paragraphs",
-    "HINT_REFLOW"      : "Reflow Paragraphs",
-    
-    "IMAGE_DIALOG"           : "Insert Image",
-    "IMAGE_TEXT_TITLE"       : "Alternate Text",
-    "IMAGE_TEXT_PLACEHOLDER" : "displayed on hover...",
-    "IMAGE_PATH_TITLE"       : "Image Path",
-    "IMAGE_PATH_PLACEHOLDER" : "e.g. http://server/image.png",
-    
-    "LINK_DIALOG"           : "Insert Link",
-    "LINK_TEXT_TITLE"       : "Link Text",
-    "LINK_TEXT_PLACEHOLDER" : "...",
-    "LINK_PATH_TITLE"       : "URL",
-    "LINK_PATH_PLACEHOLDER" : "relative or absolute...",
-
-    "SELECT_FILE" : "Select File"
-});
