--- src/extensions/default/Phoenix-markdown-toolbar/package.json
@@ -1,20 +1,0 @@
-{
-    "name": "alanhohn.markdown-toolbar",
-    "title": "Markdown Toolbar",
-    "version": "1.2.0",
-    "engines": { "brackets": ">=1.0.0" },
-    "description": "Toolbar for writing documents in Markdown.",
-    "keywords": ["markdown", "writing"],
-    "homepage": "https://github.com/alanhohn/markdown-toolbar",
-    "bugs": "https://github.com/alanhohn/markdown-toolbar/issues",
-    "license": "MIT",
-    "author": {
-        "name": "Alan Hohn",
-        "email": "alanhohn@gmail.com",
-        "url": "https://github.com/alanhohn"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/alanhohn/markdown-toolbar.git"
-    }
-}
