--- src/extensions/default/Phoenix-markdownPreview/.gitignore
@@ -1,1 +1,0 @@
-.DS_Store
